Fix bold privacy table
We had Nunito Sans as our main font but we only downloaded the 800
weight and overrode with sans for the body.

Flip this; start with sans-serif and only override to the weighty
Nunito Sans for headers.

Use that on the rights table's th elements, and add 2rem of spacing for
consistent rhythm.